pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Automators Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.47.0",
    "pytest>=8.3.3",
    "pytest-playwright>=0.5.2",
    "openai>=1.51.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.2",
]
